+++ internal/file/proto_set_provider.go
@@ -25,7 +25,6 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/uber/prototool/internal/settings"
@@ -110,7 +109,6 @@ func (c *protoSetProvider) GetMultipleForDir(workDirPath string, dirPath string)
 	if configFilePath != "" {
 		configDirPath = filepath.Dir(configFilePath)
 	}
-
 	protoFiles, err := c.walkAndGetAllProtoFiles(workDirPath, configDirPath)
 	if err != nil {
 		return nil, err
@@ -185,19 +183,18 @@ func (c *protoSetProvider) getBaseProtoSets(dirPathToProtoFiles map[string][]*Pr
 	return protoSets, nil
 }
 
-func (c *protoSetProvider) walkAndGetAllProtoFiles(workDirPath string, dirPath string) ([]*ProtoFile, error) {
-	var protoFiles []*ProtoFile
-	absWorkDirPath, err := AbsClean(workDirPath)
-	if err != nil {
-		return nil, err
-	}
-	absDirPath, err := AbsClean(dirPath)
-	if err != nil {
-		return nil, err
-	}
-	allExcludePrefixes := make(map[string]struct{})
-	numWalkedFiles := 0
-	timedOut := false
+// walkAndGetAllProtoFiles collects the .proto files nested under the given absDirPath.
+// absDirPath represents the absolute path at which the prototool.yaml configuration is
+// found, whereas absWorkDirPath represents absolute path at which prototool was invoked.
+// absWorkDirPath is only used to determine the ProtoFile.DisplayPath, also known as
+// the relative path from where prototool was invoked.
+func (c *protoSetProvider) walkAndGetAllProtoFiles(absWorkDirPath string, absDirPath string) ([]*ProtoFile, error) {
+	var (
+		protoFiles     []*ProtoFile
+		numWalkedFiles int
+		timedOut       bool
+	)
+	allExcludes := make(map[string]struct{})
 	walkErrC := make(chan error)
 	go func() {
 		walkErrC <- filepath.Walk(
@@ -212,41 +209,36 @@ func (c *protoSetProvider) walkAndGetAllProtoFiles(workDirPath string, dirPath s
 						"timed out after %v and having seen %d files, are you sure you are operating "+
 						"in the right context?", c.walkTimeout, numWalkedFiles)
 				}
-				absFilePath, err := AbsClean(filePath)
-				if err != nil {
-					return err
-				}
+				// Verify if we should skip this directory/file.
 				if fileInfo.IsDir() {
-					excludePrefixes, err := c.configProvider.GetExcludePrefixesForDir(absFilePath)
+					// Add the excluded files with respect to the current file path.
+					excludes, err := c.configProvider.GetExcludePrefixesForDir(filePath)
 					if err != nil {
 						return err
 					}
-					for _, excludePrefix := range excludePrefixes {
-						allExcludePrefixes[excludePrefix] = struct{}{}
+					for _, exclude := range excludes {
+						allExcludes[exclude] = struct{}{}
 					}
-					for excludePrefix := range allExcludePrefixes {
-						if strings.HasPrefix(absFilePath, excludePrefix) {
-							return filepath.SkipDir
-						}
+					if isExcluded(filePath, absDirPath, allExcludes) {
+						return filepath.SkipDir
 					}
 					return nil
 				}
 				if filepath.Ext(filePath) != ".proto" {
 					return nil
 				}
-				for excludePrefix := range allExcludePrefixes {
-					if strings.HasPrefix(absFilePath, excludePrefix) {
-						return nil
-					}
+				if isExcluded(filePath, absDirPath, allExcludes) {
+					return nil
 				}
+
+				// Visit this file.
 				displayPath, err := filepath.Rel(absWorkDirPath, filePath)
 				if err != nil {
-					//return err
 					displayPath = filePath
 				}
 				displayPath = filepath.Clean(displayPath)
 				protoFiles = append(protoFiles, &ProtoFile{
-					Path:        absFilePath,
+					Path:        filePath,
 					DisplayPath: displayPath,
 				})
 				return nil
@@ -270,8 +262,7 @@ func (c *protoSetProvider) walkAndGetAllProtoFiles(workDirPath string, dirPath s
 		if walkErr := <-walkErrC; walkErr != nil {
 			return nil, walkErr
 		}
-		// TODO(pedge): nice job with your code, I should learn how to code
-		return nil, fmt.Errorf("should never get here")
+		return nil, fmt.Errorf("internal prototool error")
 	}
 }
 
@@ -297,4 +288,33 @@ func getProtoFiles(filePaths []string) ([]*ProtoFile, error) {
 		})
 	}
 	return protoFiles, nil
+}
+
+// isExcluded determines whether the given filePath should be excluded.
+// Note that all excludes are assumed to be cleaned absolute paths at
+// this point.
+// stopPath represents the absolute path to the prototool configuration.
+// This is used to determine when we should stop checking for excludes.
+func isExcluded(filePath, stopPath string, excludes map[string]struct{}) bool {
+	// Use the root as a fallback so that we don't loop forever.
+	root := filepath.Dir(string(filepath.Separator))
+
+	isNested := func(curr, exclude string) bool {
+		for {
+			if curr == stopPath || curr == root {
+				return false
+			}
+			if curr == exclude {
+				return true
+			}
+			curr = filepath.Dir(curr)
+		}
+	}
+	for exclude := range excludes {
+		if isNested(filePath, exclude) {
+			return true
+		}
+	}
+	return false
+
 }