+++ internal/settings/config_provider.go
@@ -127,7 +127,7 @@ func get(filePath string) (Config, error) {
 //
 // This will return a valid Config, or an error.
 func externalConfigToConfig(e ExternalConfig, dirPath string) (Config, error) {
-	excludePrefixes, err := getExcludePrefixes(e.Excludes, true /* Never include default excludes */, dirPath)
+	excludePrefixes, err := getExcludePrefixes(e.Excludes, dirPath)
 	if err != nil {
 		return Config{}, err
 	}
@@ -163,12 +163,6 @@ func externalConfigToConfig(e ExternalConfig, dirPath string) (Config, error) {
 		if plugin.Output == "" {
 			return Config{}, fmt.Errorf("output path required for plugin %s", plugin.Name)
 		}
-		path := ""
-		if len(e.Gen.PluginOverrides) > 0 {
-			if override, ok := e.Gen.PluginOverrides[plugin.Name]; ok && override != "" {
-				path = override
-			}
-		}
 		var relPath, absPath string
 		if filepath.IsAbs(plugin.Output) {
 			absPath = filepath.Clean(plugin.Output)
@@ -182,7 +176,7 @@ func externalConfigToConfig(e ExternalConfig, dirPath string) (Config, error) {
 		}
 		genPlugins[i] = GenPlugin{
 			Name:  plugin.Name,
-			Path:  path,
+			Path:  plugin.Path,
 			Type:  genPluginType,
 			Flags: plugin.Flags,
 			OutputPath: OutputPath{
@@ -272,30 +266,22 @@ func externalConfigToConfig(e ExternalConfig, dirPath string) (Config, error) {
 func getExcludePrefixesForDir(dirPath string) ([]string, error) {
 	filePath := filepath.Join(dirPath, DefaultConfigFilename)
 	if _, err := os.Stat(filePath); err != nil {
-		excludePrefixes := make([]string, 0, len(DefaultExcludePrefixes))
-		for _, defaultExcludePrefix := range DefaultExcludePrefixes {
-			excludePrefixes = append(excludePrefixes, filepath.Join(dirPath, defaultExcludePrefix))
-		}
-		return excludePrefixes, nil
+		return []string{}, nil
 	}
 	data, err := ioutil.ReadFile(filePath)
 	if err != nil {
 		return nil, err
 	}
 	s := struct {
-		ExcludePaths          []string `json:"excludes,omitempty" yaml:"excludes,omitempty"`
-		NoDefaultExcludePaths bool     `json:"no_default_excludes,omitempty" yaml:"no_default_excludes,omitempty"`
+		ExcludePaths []string `json:"excludes,omitempty" yaml:"excludes,omitempty"`
 	}{}
 	if err := yaml.Unmarshal(data, &s); err != nil {
 		return nil, err
 	}
-	return getExcludePrefixes(s.ExcludePaths, s.NoDefaultExcludePaths, dirPath)
+	return getExcludePrefixes(s.ExcludePaths, dirPath)
 }
 
-func getExcludePrefixes(excludes []string, noDefaultExcludes bool, dirPath string) ([]string, error) {
-	if !noDefaultExcludes {
-		excludes = append(DefaultExcludePrefixes, excludes...)
-	}
+func getExcludePrefixes(excludes []string, dirPath string) ([]string, error) {
 	excludePrefixes := make([]string, 0, len(excludes))
 	for _, excludePrefix := range strs.DedupeSort(excludes, nil) {
 		if !filepath.IsAbs(excludePrefix) {